fix relative import and test structure
