+++ trackintel/io/__init__.py
@@ -1,3 +1,24 @@
-from .file import *
-from .postgis import *
-from .dataset_reader import *
+from .file import read_positionfixes_csv
+from .file import write_positionfixes_csv
+from .postgis import read_positionfixes_postgis
+from .postgis import write_positionfixes_postgis
+
+from .file import read_triplegs_csv
+from .file import write_triplegs_csv
+from .postgis import read_triplegs_postgis
+from .postgis import write_triplegs_postgis
+
+from .file import read_staypoints_csv
+from .file import write_staypoints_csv
+from .postgis import read_staypoints_postgis
+from .postgis import write_staypoints_postgis
+
+from .file import read_locations_csv
+from .file import write_locations_csv
+from .postgis import read_locations_postgis
+from .postgis import write_locations_postgis
+
+from .file import read_trips_csv
+from .file import write_trips_csv
+from .postgis import read_trips_postgis
+from .postgis import write_trips_postgis
