+++ trackintel/preprocessing/__init__.py
@@ -1,4 +1,10 @@
-from .positionfixes import *
-from .staypoints import *
-from .triplegs import *
-from .filter import *
+from .positionfixes import generate_staypoints
+from .positionfixes import generate_triplegs
+
+from .filter import spatial_filter
+
+from .staypoints import generate_locations
+from .staypoints import create_activity_flag
+
+from .triplegs import smoothen_triplegs
+from .triplegs import generate_trips
