+++ trackintel/visualization/__init__.py
@@ -1,3 +1,9 @@
-from .positionfixes import *
-from .staypoints import *
-from .triplegs import *
+from .locations import plot_center_of_locations
+
+from .osm import plot_osm_streets
+
+from .positionfixes import plot_positionfixes
+
+from .staypoints import plot_staypoints
+
+from .triplegs import plot_triplegs
